Enable calculation of multiple metrics for the same structure.
